fix(server): mint per-game JSON log rows at the session layer
Review found the write hook lived at one phase-server transport call
site and silently missed five other production paths that also
produce GameLogEntry rows via GameSession::run_ai (takeback approve/
respond, in-game reconnect, game start, Play-vs-AI start) — all call
run_ai directly with no wrapper, so nothing wrote their entries.

Move GameFileCache into server-core (new game_log module, Stream enum
replacing the stringly-typed stream key) and hook it at the three
functions that actually mint ActionResults: SessionManager's
handle_action/handle_interaction_with_rejection, and GameSession's
run_ai_action_batch — the single funnel every run_ai() caller goes
through. Threaded the same way GameSession.hosting already is
(stamped at creation, re-stamped at restore, never deserialized).

Writing at the mint point instead of after a transport-layer guard
also fixes two review findings for free: entries no longer vanish
when the post-action broadcast-size guard rejects a snapshot (the
write now happens first), and ordering is no longer racy across
concurrent socket tasks (the write happens while the session lock is
held). Also: test fixtures now match production's always-zero
GameLogEntry.seq instead of asserting on invented values, the
disabled-cache test asserts on the writer map instead of a
constructor property, and reopen-after-close/bad-path-sentinel
coverage that the first pass dropped is back.
